refactor: move useImFeelingLucky to its own file
